Add AppTemplate subfolder and move FormStateAutoSaver there, in order to facilitate development of other applications using the same template.
